move page into pages folder
